+++ src/test/java/ru/bank/products/DebitCardTest.java
@@ -1,6 +1,7 @@
 package ru.bank.products;
 
 import io.qameta.allure.Description;
+import java.math.BigDecimal;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 
@@ -9,15 +10,15 @@ public class DebitCardTest {
     @Test
     @Description("Проверка корректного пополнения баланса дебетовой карты")
     public void testDebitCardDeposit() {
-        DebitCard card = new DebitCard("RUB", 1000, "Дебетовая карта");
-        card.deposit(500);
-        Assert.assertEquals(card.getBalance(), 1500);
+        DebitCard card = new DebitCard("RUB", BigDecimal.valueOf(1000), "Дебетовая карта");
+        card.deposit(BigDecimal.valueOf(500));
+        Assert.assertEquals(card.getBalance(), BigDecimal.valueOf(1500));
     }
 
     @Test(expectedExceptions = IllegalArgumentException.class)
     @Description("Проверка исключения при попытке списания суммы больше доступного баланса на дебетовой карте")
     public void testDebitCardWithdrawException() {
-        DebitCard card = new DebitCard("RUB", 1000, "Дебетовая карта");
-        card.withdraw(2000);
+        DebitCard card = new DebitCard("RUB", BigDecimal.valueOf(1000), "Дебетовая карта");
+        card.withdraw(BigDecimal.valueOf(2000));
     }
 }
